+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # node-red-contrib-sun-position
 
+#### 1.0.13:  BugFix
+
+- blind-control + clock-time + time-inject + within-time
+  - fixed bug where default expire time is not considered #112
+  - reduced usage of context for store data
+
+- general
+  - i18N
+
 #### 1.0.12:  BugFix + enhancement
+
+##### fixes
 
 - general
   - BugFix: Allow 0 as value for any time input, mainly for time-compare or time-span node, but affects potentially all nodes.
@@ -8,10 +19,10 @@
   - BugFix: node error output for time input if time can not be evaluated has missing original message.
   - prepared new Node interval-inject (not finished, not available) - is the same as standard node-red inject node for intervals with start and end of interval able to use sun-times.
 
-enhancements
+##### enhancements
 
-  - Blind-control allows now granular settings of topic
-  - documentation enhanced for clock-timer and general
+- Blind-control allows now granular settings of topic
+- documentation enhanced for clock-timer and general
 
 #### 1.0.11:  enhancement